symsrv: remove default cache location logic
The logic involving searching and store to a default cache location is
quite complicated. The rules for when this location is used depend on if
the user specified a cache location. It is much easier to assume no
cache location if the user did not specify one. It is a rare occurrence
that a user would intentionally omit the cache path in order to use the
default.